Add FluentValidation, Logger, some other minor changes

+++ UserHouse.Domain/Users/UserAppService.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using AutoMapper;
 using Microsoft.EntityFrameworkCore.Migrations.Operations;
+using Microsoft.Extensions.Logging;
 using UserHouse.Application.Models;
 using UserHouse.Data.Entities;
 using UserHouse.Data.Repositories.Users;
@@ -14,13 +15,16 @@ namespace UserHouse.Application.Users
 {
     public class UserAppService : IUserAppService
     {
+        private readonly ILogger<UserAppService> _logger;
         private readonly IGenericRepository<User> _userRepository;
         private readonly IMapper _mapper;
 
         public UserAppService(
+            ILogger<UserAppService> logger,
             IGenericRepository<User> userRepository,
             IMapper mapper)
         {
+            _logger = logger;
             _userRepository = userRepository;
             _mapper = mapper;
         }
@@ -29,6 +33,12 @@ namespace UserHouse.Application.Users
         {
             var users = _userRepository.GetAll();
 
+            if (users == null)
+            {
+                _logger.LogInformation("List of users is empty");
+                throw new Exception("Unable to get users!");
+            }
+
             return _mapper.Map<List<UserModel>>(users);
         }
 
@@ -36,35 +46,54 @@ namespace UserHouse.Application.Users
         {
             var user = await _userRepository.Get(userId);
 
+            if (user == null)
+            {
+                _logger.LogInformation($"Unable to find user with Id: {userId}");
+                throw new Exception("Unable to get specified user!");
+            }
+
             return _mapper.Map<UserModel>(user);
         }
 
-        public void Create(UserModel createUserDto)
+        public void Create(UserModel userModel)
         {
-            var newUser = _mapper.Map<User>(createUserDto);
+            if (userModel == null)
+            {
+                _logger.LogInformation("Unable to create a new user due to empty data in UserModel");
+                throw new Exception("Unable to create a new user!");
+            }
+
+            var newUser = _mapper.Map<User>(userModel);
 
             newUser.DateOfBirth = DateTime.Now;
 
             _userRepository.Add(newUser);
         }
 
-        public void Update(UserModel userDto)
+        public void Update(UserModel userModel)
         {
-            var updatedUser = _mapper.Map<User>(userDto);
+            if (userModel == null)
+            {
+                _logger.LogInformation("Unable to update user due to empty data in UserModel");
+                throw new Exception("Unable to update specified user!");
+            }
+
+            var updatedUser = _mapper.Map<User>(userModel);
 
             _userRepository.Update(updatedUser);
         }
 
-        public void Delete(UserModel userModel)
+        public async void Delete(int userId)
         {
-            var user = _mapper.Map<User>(userModel);
+            var user = await _userRepository.Get(userId);
+
+            if (user == null)
+            {
+                _logger.LogInformation($"Unable to find user with Id: {userId} and delete it");
+                throw new Exception("Unable to delete specified user!");
+            }
 
             _userRepository.Delete(user);
-        }
-
-        public void DoSomethingInAppService()
-        {
-            
         }
     }
 }
